Move mapSrc to util
